Some GUI improvements.
http://developer.gnome.org/projects/gup/hig/2.0/menus-design.html

Label the menu item with a trailing ellipsis ("...") only if the command
requires further input from the user before it can be performed. Do not
add an ellipsis to items that only present a confirmation dialog (such
as Delete), or that do not require further input (such as Properties,
Preferences or About).